fix: Add external data file for charts & progress bar in admin panel
to show the progress for the target set by company to hire 20000 candidates from India & monthly hires in charts
