+++ jobseekers.com/src/Components/admin-panel/adminComponents/chart/Chart.jsx
@@ -1,15 +1,7 @@
 import './chart.scss'
-import { AreaChart, Area, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts'
-  ;
+import { AreaChart, Area, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts';
+import data from './totalHiredDtata';
 
-const data = [
-{name: "January", Total: 1200},
-{name: "February", Total: 2100},
-{name: "March", Total: 800},
-{name: "April", Total: 1600},
-{name: "May", Total: 900},
-{name: "June", Total: 1700},
-];
 const Chart = () => {
   return (
     <div className='chart'>
