+++ jobseekers.com/src/Components/admin-panel/adminComponents/chart/totalHiredDtata.js
@@ -1,0 +1,11 @@
+
+const data = [
+    {name: "January", Total: 1200},
+    {name: "February", Total: 2100},
+    {name: "March", Total: 800},
+    {name: "April", Total: 1600},
+    {name: "May", Total: 900},
+    {name: "June", Total: 1700},
+    ];
+
+export default data;
