+++ jobseekers.com/src/Components/admin-panel/adminComponents/featured/pastData.js
@@ -1,0 +1,13 @@
+const yearlyData=[
+    {
+        id: 1,
+        target: "20,000",
+        lastEightMonthsHires: "14000",
+        ios: "4900",
+        uiux: "3100",
+        react: "2500",
+        swift: "3500"
+    }
+]
+
+export default yearlyData;
